"Doublestrike" to "Double Strike"
The book is inconsistent with formatting for "Double Strike" vs "Doublestrike", so inevitably it shows up in the data, and can obviously be annoying to deal with in practice. If this is from my data, I apologize. Must've forgotten to edit that particular issue. Since "Five Strike" is consistently that, "Double Strike" makes sense.